+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ### Version 2.0.9 ([Maven Central](http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.0.9%22)) ###
 
-* 
+* update truncate_html filter to support preserving words by default, with an additional parameter to chop words at length
 
 ### Version 2.0.8 ([Maven Central](http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.hubspot.jinjava%22%20AND%20v%3A%222.0.8%22)) ###